flatten: nokhte widget hierarchy
