+++ packages/client/lib/app/core/widgets/nokhte/canvas/nokhte_painter.dart
