--- packages/client/lib/app/core/widgets/nokhte/stack/presentation/constants/constants.dart
@@ -1,2 +1,0 @@
-export 'movies/movies.dart';
-export 'types/types.dart';
