--- packages/client/lib/app/core/widgets/nokhte/stack/presentation/presentation.dart
@@ -1,3 +1,0 @@
-export 'mobx/nokhte_store.dart';
-export 'widget/nokhte.dart';
-export 'constants/constants.dart';
